2014-07-05 Roman - Archive: TArchive, TMArchivator and TMArchivator moved to use TConfig::cntrCmdMake(). - FIX:Transport: TTransportOut, TTransportIn fix for sets values check and process. - !!:SYS: TConfig::cfgChange() expanded by previous configuration value argument "pc".     Library version set to 3:2:0. - UI.QTCfg: Selected group nodes commands "load" and "save" support is added.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2128 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tvariant.cpp
@@ -958,7 +958,7 @@ AutoHD<XMLNodeObj> XMLNodeObj::childGet( const string &name, unsigned num )
     oRes.resRequestR();
     AutoHD<XMLNodeObj> rez;
     for(int i_ch = 0, i_n = 0; i_ch < (int)mChilds.size(); i_ch++)
-	if(strcasecmp(mChilds[i_ch].at().name().c_str(),name.c_str()) == 0 && i_n++ == num)
+	if(strcasecmp(mChilds[i_ch].at().name().c_str(),name.c_str()) == 0 && (i_n++) == (int)num)
 	    rez = mChilds[i_ch];
     oRes.resRelease();
     if(rez.freeStat()) throw TError("XMLNodeObj",_("Child %s:%d is not found!"),name.c_str(),num);
